% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{plot_dens_boundaries}
\alias{plot_dens_boundaries}
\title{Plot Density curve with colored area inside given boundaries}
\usage{
plot_dens_boundaries(
  samples,
  boundaries,
  posterior = NULL,
  xlabel = "posterior",
  ylabel = "density",
  title = ""
)
}
\arguments{
\item{samples}{vector containing samples drawn from the posterior}

\item{boundaries}{lower or upper boundaries for coloring the area below
the density}

\item{posterior}{(optional) vector containing the posterior}

\item{xlabel, ylabel, title}{strings for formating the plot}
}
\value{
plot
}
\description{
Plot Density curve with colored area inside given boundaries
}
