% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{plot_grid_hdpi}
\alias{plot_grid_hdpi}
\title{Plot Density curve with colored area inside the HPDI for the given
probability, having sampled from the posterior.}
\usage{
plot_grid_hdpi(posterior, prob, sample_size = 10000, xlabel = "probability")
}
\arguments{
\item{posterior}{vector containing the posterior}

\item{prob}{probability density}

\item{sample_size}{number of samples to be drawn from the posterior}

\item{xlabel}{string for formating the plot}
}
\value{
plot
}
\description{
Plot Density curve with colored area inside the HPDI for the given
probability, having sampled from the posterior.
}
