% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgplot-package.R
\docType{package}
\name{rgplot-package}
\alias{rgplot}
\alias{rgplot-package}
\title{rgplot: Statistical Rethinking ggplot utils}
\description{
Provides a series of utility functions for plotting with `ggplot` the objects produced while working on McElreath's Statistical Rethinking.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caiogeraldes/rgplot}
  \item Report bugs at \url{https://github.com/caiogeraldes/rgplot/issues}
}

}
\author{
\strong{Maintainer}: Caio Geraldes \email{caio.geraldes@usp.br}

}
\keyword{internal}
